+++ scripts/modeldb.sh
@@ -20,25 +20,31 @@ dbMenu() {
             do
                 if [ $create -eq 0 ]; then
                     read -p "Do you want to create another Database (y/n) " answer
-                    answer=$(echo $answer | tr '[:upper:]' '[:lower:]')
-                    if [ $answer = "y" ] || [ $answer = "yes" ]; then
-                            create=$($scriptsPath/createdb.sh)
-                    elif [ $answer = "n" -o $answer = "no" ]; then
-                        break
-                    else
-                        echo "$red$bg please chose from this values only (y/n).$end"
-                    fi
-                elif [ $create -eq 1 ]; then
-                    read -p "Do you still want to create a Database? (y/n) " answer
-                    answer=$(echo $answer | tr '[:upper:]' '[:lower:]')
-                    check=$($scriptsPath/chkname.sh $answer)
-                    if [ $check -eq 0 ]; then
+                    chechAnswer=$($scriptsPath/chkname.sh $answer)
+                    if [ $chechAnswer -eq 0 ]; then
+                        answer=$(echo $answer | tr '[:upper:]' '[:lower:]')
                         if [ $answer = "y" ] || [ $answer = "yes" ]; then
                                 create=$($scriptsPath/createdb.sh)
                         elif [ $answer = "n" -o $answer = "no" ]; then
                             break
                         else
-                            echo "$red$bg please choose from this values (y/n).$end"
+                            echo "$red$bg please chose from this values only (y/n).$end"
+                        fi
+                    fi
+                elif [ $create -eq 1 ]; then
+                    read -p "Do you still want to create a Database? (y/n) " answer
+                    chechAnswer=$($scriptsPath/chkname.sh $answer)
+                    if [ $chechAnswer -eq 0 ]; then
+                        answer=$(echo $answer | tr '[:upper:]' '[:lower:]')
+                        check=$($scriptsPath/chkname.sh $answer)
+                        if [ $check -eq 0 ]; then
+                            if [ $answer = "y" ] || [ $answer = "yes" ]; then
+                                    create=$($scriptsPath/createdb.sh)
+                            elif [ $answer = "n" -o $answer = "no" ]; then
+                                break
+                            else
+                                echo "$red please choose from this values (y/n).$end"
+                            fi
                         fi
                     fi
                 fi
@@ -59,25 +65,31 @@ dbMenu() {
             do
                 if [ $dropdb -eq 0 ]; then
                     read -p "Do you want to Drop another Database? (y/n)? " answer
-                    answer=$(echo $answer | tr '[:upper:]' '[:lower:]')
-                    if [ $answer = "y" ] || [ $answer = "yes" ]; then
-                            dropdb=$($scriptsPath/dropdb.sh)
-                    elif [ $answer = "n" -o $answer = "no" ]; then
-                        break
-                    else
-                        echo "$red$bg please choose from this values (y/n).$end"
-                    fi
-                elif [ $dropdb -eq 1 ]; then
-                    read -p "Do you still want to Drop a Database? (y/n)? " answer
-                    answer=$(echo $answer | tr '[:upper:]' '[:lower:]')
-                    check=$($scriptsPath/chkname.sh $answer)
-                    if  [ $check -eq 0 ]; then
-                        if [ $answer = "y" -o $answer = "yes" ]; then
-                                create=$($scriptsPath/dropdb.sh)
+                    chechAnswer=$($scriptsPath/chkname.sh $answer)
+                    if [ $chechAnswer -eq 0 ]; then
+                        answer=$(echo $answer | tr '[:upper:]' '[:lower:]')
+                        if [ $answer = "y" ] || [ $answer = "yes" ]; then
+                                dropdb=$($scriptsPath/dropdb.sh)
                         elif [ $answer = "n" -o $answer = "no" ]; then
                             break
                         else
-                            echo "${red}Please choose from this values(y/n).$end"
+                            echo "$red please choose from this values (y/n).$end"
+                        fi
+                    fi
+                elif [ $dropdb -eq 1 ]; then
+                    read -p "Do you still want to Drop a Database? (y/n)? " answer
+                    chechAnswer=$($scriptsPath/chkname.sh $answer)
+                    if [ $chechAnswer -eq 0 ]; then
+                        answer=$(echo $answer | tr '[:upper:]' '[:lower:]')
+                        check=$($scriptsPath/chkname.sh $answer)
+                        if  [ $check -eq 0 ]; then
+                            if [ $answer = "y" -o $answer = "yes" ]; then
+                                    create=$($scriptsPath/dropdb.sh)
+                            elif [ $answer = "n" -o $answer = "no" ]; then
+                                break
+                            else
+                                echo "${red}Please choose from this values(y/n).$end"
+                            fi
                         fi
                     fi
                 fi
